% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstd.gen.R
\name{bundled_cli_vrsn}
\alias{bundled_cli_vrsn}
\title{Determine version of CLI tool bundled with RStudio}
\usage{
bundled_cli_vrsn(tool = all_bundled_tools)
}
\arguments{
\item{tool}{Tool name. One of \code{"dart-sass"}, \code{"deno"}, \code{"esbuild"}, \code{"pandoc"} or \code{"quarto"}.}
}
\value{
A \link[=numeric_version]{numeric version}.
}
\description{
Determines the version of one of the command-line interface (CLI) tools bundled with RStudio, like \href{https://quarto.org/}{Quarto},
\href{https://pandoc.org/}{Pandoc}, \href{https://sass-lang.com/dart-sass/}{Dart Sass}, etc.
}
\examples{
rstd::bundled_cli_vrsn(tool = "dart-sass")
}
