% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstd.gen.R
\name{bundled_pandoc_version}
\alias{bundled_pandoc_version}
\title{Determine version of Pandoc bundled with RStudio}
\usage{
bundled_pandoc_version()
}
\value{
A \link[=numeric_version]{numeric version}.
}
\description{
Determine version of Pandoc bundled with RStudio
}
