% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstd.gen.R
\name{is_uptodate}
\alias{is_uptodate}
\title{Test if RStudio is up to date}
\usage{
is_uptodate(stable = TRUE, os = NULL)
}
\arguments{
\item{stable}{Set to \code{FALSE} in order to test against the latest \href{https://rstudio.com/products/rstudio/download/preview/}{RStudio preview build} instead
of the latest \href{https://rstudio.com/products/rstudio/download/}{stable build}.}

\item{os}{The OS \emph{codename} for which the RStudio version was built. If \code{NULL}, it will be auto-detected for the current system.}
}
\value{
\code{TRUE} if the currently running RStudio version is the latest version, \code{FALSE} otherwise.
}
\description{
Test if RStudio is up to date
}
