% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstd-GEN.R
\name{latest_version}
\alias{latest_version}
\title{Get latest RStudio version number}
\usage{
latest_version(
  type = c("desktop", "server"),
  stable = TRUE,
  pro = FALSE,
  os = NULL,
  use_cache = TRUE,
  cache_lifespan = "1 day"
)
}
\arguments{
\item{type}{Either \code{"desktop"} for \href{https://rstudio.com/products/rstudio/#rstudio-desktop}{RStudio Desktop} or \code{"server"} for
\href{https://rstudio.com/products/rstudio/#rstudio-server}{RStudio Server} release metadata.}

\item{stable}{Set to \code{FALSE} to retrieve release metadata of \href{https://rstudio.com/products/rstudio/download/preview/}{RStudio preview builds} instead of
\href{https://rstudio.com/products/rstudio/download/}{stable builds}.}

\item{pro}{\code{TRUE} for the proprietary RStudio (Server) Pro edition and \code{FALSE} for the open-source RStudio (Server) edition.}

\item{os}{The OS \emph{codename} for which the RStudio version was built. If \code{NULL}, it will be auto-detected for the current system.}

\item{use_cache}{Return cached results if possible. If \code{FALSE}, results are newly fetched (and at the same time the cache gets refreshed).}

\item{cache_lifespan}{The duration after which cached results are refreshed (i.e. newly fetched). A valid \link[lubridate:as.duration]{lubridate duration}.
Defaults to 1 day (24 hours). Only relevant if \code{use_cache = TRUE}.}
}
\value{
A \link[=numeric_version]{numeric version}.
}
\description{
Get latest RStudio version number
}
\examples{
latest_version(os = "macos",
               use_cache = FALSE)
}
