% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstd.gen.R
\docType{data}
\name{pkg_config}
\alias{pkg_config}
\title{rstd package configuration metadata}
\format{
A \link[tibble:tbl_df-class]{tibble} with the columns \code{key}, \code{default_value} and \code{description}.
}
\usage{
pkg_config
}
\description{
A \link[tibble:tbl_df-class]{tibble} with metadata of all possible rstd package configuration options. See \code{\link[pal:pkg_config_val]{pal::pkg_config_val()}} for more information.
}
\examples{
rstd::pkg_config
}
\keyword{datasets}
