% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstd.gen.R
\name{pkg_status}
\alias{pkg_status}
\title{List RStudio's R package dependencies' installation status}
\usage{
pkg_status()
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
List RStudio's R package dependencies' installation status
}
