% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstd.gen.R
\name{releases}
\alias{releases}
\title{Get RStudio release metadata}
\usage{
releases(
  type = c("desktop", "server"),
  stable = TRUE,
  use_cache = TRUE,
  max_cache_age = "1 day"
)
}
\arguments{
\item{type}{Either \code{"desktop"} for \href{https://rstudio.com/products/rstudio/#rstudio-desktop}{RStudio Desktop} or \code{"server"} for
\href{https://rstudio.com/products/rstudio/#rstudio-server}{RStudio Server} release metadata.}

\item{stable}{Set to \code{FALSE} to retrieve release metadata of \href{https://rstudio.com/products/rstudio/download/preview/}{RStudio preview builds} instead of
\href{https://rstudio.com/products/rstudio/download/}{stable builds}.}

\item{use_cache}{Return cached results if possible. If \code{FALSE}, results are always newly fetched regardless of \code{max_cache_age}.}

\item{max_cache_age}{The duration after which cached results are refreshed (i.e. newly fetched). A valid \link[lubridate:as.duration]{lubridate duration}. Use \code{Inf} to disable cache expiry. Only relevant if \code{use_cache = TRUE}. Defaults to 1 day (24 hours).}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Get RStudio release metadata
}
\examples{
releases(type = "server",
         max_cache_age = "1 year 2 months 3 weeks 4 days 5 hours 6 minutes 7 seconds")
}
