% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstd-package.R
\docType{package}
\name{rstd-package}
\alias{rstd}
\alias{rstd-package}
\title{rstd: Unofficial Utility Functions Around the RStudio IDE}
\description{
Get RStudio Desktop release metadata and bundled Pandoc version. Test if RStudio Desktop is up to date.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rstd.rpkg.dev}
  \item \url{https://gitlab.com/salim_b/r/pkgs/rstd}
  \item Report bugs at \url{https://gitlab.com/salim_b/r/pkgs/rstd/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
