% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstd-package.R
\docType{package}
\name{rstd-package}
\alias{rstd}
\alias{rstd-package}
\title{rstd: Unofficial Utility Functions Around the RStudio IDE}
\description{
Get RStudio Desktop release metadata and bundled Pandoc version. Test if RStudio Desktop is up to date.
}
\section{Package configuration}{


Some of rstd's functionality is controlled via package-specific global configuration which can either be set via \href{https://rdrr.io/r/base/options.html}{R options} or \href{https://en.wikipedia.org/wiki/Environment_variable}{environment variables} (the former take precedence). This configuration includes:\tabular{llll}{
   \strong{Description} \tab \strong{R option} \tab \strong{Environment variable} \tab \strong{Default value} \cr
   Maximal timespan to preserve the package's \href{https://pkgpins.rpkg.dev/}{pkgpins} cache. Cache entries older than this will be deleted upon package loading. \tab \code{rstd.global_max_cache_age} \tab \code{RSTD_GLOBAL_MAX_CACHE_AGE} \tab \code{"30 days"} \cr
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://rstd.rpkg.dev}
  \item \url{https://gitlab.com/rpkg.dev/rstd}
  \item Report bugs at \url{https://gitlab.com/rpkg.dev/rstd/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
